+++ wallet/src/message_interface/message.rs
@@ -23,9 +23,9 @@ use crate::{
 #[serde(tag = "cmd", content = "payload", rename_all = "camelCase")]
 #[allow(clippy::large_enum_variant)]
 pub enum Message {
-    ClientMessage {
-        message: ClientMessage,
-    },
+    /// Call the client message interface.
+    /// Expected response: [`Client`](crate::message_interface::Response::Client)
+    Client { message: ClientMessage },
     /// Creates an account.
     /// Expected response: [`Account`](crate::message_interface::Response::Account)
     CreateAccount {
@@ -126,9 +126,7 @@ pub enum Message {
     GenerateMnemonic,
     /// Checks if the given mnemonic is valid.
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
-    VerifyMnemonic {
-        mnemonic: String,
-    },
+    VerifyMnemonic { mnemonic: String },
     /// Updates the client options for all accounts.
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
     SetClientOptions {
@@ -169,9 +167,7 @@ pub enum Message {
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
     #[cfg(feature = "stronghold")]
     #[cfg_attr(docsrs, doc(cfg(feature = "stronghold")))]
-    SetStrongholdPassword {
-        password: String,
-    },
+    SetStrongholdPassword { password: String },
     /// Set the stronghold password clear interval.
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
     #[cfg(feature = "stronghold")]
@@ -184,9 +180,7 @@ pub enum Message {
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
     #[cfg(feature = "stronghold")]
     #[cfg_attr(docsrs, doc(cfg(feature = "stronghold")))]
-    StoreMnemonic {
-        mnemonic: String,
-    },
+    StoreMnemonic { mnemonic: String },
     /// Start background syncing.
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
     StartBackgroundSync {
@@ -203,9 +197,7 @@ pub enum Message {
     /// Expected response: [`Ok`](crate::message_interface::Response::Ok)
     #[cfg(feature = "events")]
     #[cfg_attr(docsrs, doc(cfg(feature = "events")))]
-    EmitTestEvent {
-        event: WalletEvent,
-    },
+    EmitTestEvent { event: WalletEvent },
     /// Transforms a bech32 encoded address to hex
     /// Expected response: [`HexAddress`](crate::message_interface::Response::HexAddress)
     Bech32ToHex {
@@ -244,7 +236,7 @@ impl Debug for Message {
     fn fmt(&self, f: &mut Formatter<'_>) -> Result {
         match self {
             Self::ClientMessage { message } => {
-                write!(f, "ClientMessage {{ message: {message:?} }}")
+                write!(f, "ClientMessage{{ message: {message:?} }}")
             }
             Self::CreateAccount { alias, bech32_hrp } => {
                 write!(f, "CreateAccount{{ alias: {alias:?}, bech32_hrp: {bech32_hrp:?} }}")